# protobuf/swg/auth.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "auth.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/_v1/close_session": {
      "get": {
        "operationId": "CloseSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderBool"
            }
          }
        },
        "tags": [
          "ServiceAuth"
        ]
      }
    },
    "/v1/start_session": {
      "post": {
        "operationId": "StartSession",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderSession"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderAuthenticate"
            }
          }
        ],
        "tags": [
          "ServiceAuth"
        ]
      }
    }
  },
  "definitions": {
    "orderAuthenticate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "orderBool": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "orderParticipant": {
      "type": "object",
      "properties": {
        "bsonId": {
          "type": "string",
          "format": "byte",
          "title": "@inject_tag: bson:\"_id,omitempty\""
        },
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "title": "Может быть много полей"
        }
      }
    },
    "orderSession": {
      "type": "object",
      "properties": {
        "Token": {
          "type": "string"
        }
      }
    }
  }
}